research/roi_budgeting/requirements.txt
PyYAML>=6.0
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
seaborn>=0.13
scipy>=1.12
tqdm>=4.66
opencv-python-headless>=4.10
jupyterlab>=4.0
